visual changes. dragbeat changes. mapmaker changes.
mapmaker now has alot more functionality. it can now make dragbeats! you can go through individual frames and place your beats. making maps is getting alot easier, you can see all the beats on the screen as time passes on.
dragbeats are getting a massive overhaul. I've already made A TON of changes to it. the drag gets smaller the while time passes and it wont just disappear when the first point has expired. the entire dragbeat will slowly reveal itself to the player as the maptime progresses, giving more flow to it. also gives the player a look into how fast he should be moving it. 
still have to fix that you can complete the dragbeat even when its not fully visable. also will fix the bug where the dragbeat will follow with the end if the player hasent clicked it when it starts to expier.