different-dragtype
i will make a different type of drag to see if it would work better.
it will be more like a freehand drag. where you keep your mouse in the drag untill the finish and not have to drag a ball locked on a line. 
this will fix the jump from one line to the other.
this will make it feel more natural.